+++ internal/command/testdata/test/mock_provider_source_validation/main.tf
@@ -1,0 +1,5 @@
+provider "test" {}
+
+resource "test_resource" "primary" {
+  value = "foo"
+}